apps/agno/pyproject.toml
[tool.poetry]
name = "agui-agno"
version = "0.1.0"
description = ""
authors = ["orca-copilotkit <testapp@copilotkit.ai>"]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
agno = ">=1.7.3,<2.0.0"
dotenv = ">=0.9.9,<0.10.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
openai = ">=1.96.1,<2.0.0"
yfinance = ">=0.2.65,<0.3.0"
ag-ui-protocol = ">=0.1.8,<0.2.0"
numpy = ">=2.3.2,<3.0.0"
pandas = ">=2.3.1,<3.0.0"
lxml = ">=6.0.0,<7.0.0"
PyJWT = ">=2.8.0,<3.0.0"
requests = ">=2.31.0,<3.0.0"
fastmcp = ">=2.13.1"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"